Jack will only use portaudio on Windows, so make dep Windows-only.
I noticed that the linux build was detecting the presence of portaudio
but still not building support for it (and conda-build was warning that
it was a run dependency but didn't seem to be used). Inspecting the Jack
waf build script shows that they only enable portaudio support on
Windows.